refactor(employees): improve delete confirmation and formatting

+++ src/app/employees/employees.component.ts
@@ -5,25 +5,30 @@ import { MatButtonModule } from '@angular/material/button';
 import Employee from '../employee.interface';
 import { EmployeeService } from '../employee.service';
 
-
 @Component({
   selector: 'app-employees',
   standalone: true,
   imports: [CommonModule, RouterLink, MatButtonModule],
-  templateUrl: './employees.component.html'
+  templateUrl: './employees.component.html',
 })
 export class EmployeesComponent {
   employees: Employee[] = [];
   employeeService = inject(EmployeeService);
 
   ngOnInit() {
-    this.employeeService.getEmployees().subscribe(employees => this.employees = employees);
+    this.employeeService
+      .getEmployees()
+      .subscribe((employees) => (this.employees = employees));
   }
 
   deleteEmployee(id: string) {
-    const ok = confirm("Do you really want to delete this employee?");
-    this.employeeService.deleteEmployee(id).subscribe(() => {
-      this.employees = this.employees.filter(employee => employee._id !== id);
-    });
+    const ok = confirm('Do you really want to delete this employee?');
+    if (ok) {
+      this.employeeService.deleteEmployee(id).subscribe(() => {
+        this.employees = this.employees.filter(
+          (employee) => employee._id !== id
+        );
+      });
+    }
   }
 }
